backend/requirement.txt
pandas 
pyarrow 
astapi 
uvicorn 
python-multipart 
joblib 
PyPDF2